More aggressive code formatting.

+++ scripts/build-zip.mjs
@@ -1,6 +1,7 @@
 import archiver from "archiver";
 import fs from "fs";
 import path from "path";
+
 import { default as manifest } from "../system.json" with { type: "json" };
 
 const VERSION = manifest.version;
@@ -14,14 +15,11 @@ const ASSETS = ["css", "src", "packs", "lang", "system.json", "README.md"];
 
 console.log(`--- Starting Build v${VERSION} ---`);
 
-if (fs.existsSync(DIST_DIR)) {
-  fs.rmSync(DIST_DIR, { recursive: true, force: true });
-}
+if (fs.existsSync(DIST_DIR)) fs.rmSync(DIST_DIR, { force: true, recursive: true });
 fs.mkdirSync(SYSTEM_DIR, { recursive: true });
 
-for (const dst of [MANIFEST_DST, MANIFEST_VERSIONED_DST]) {
+for (const dst of [MANIFEST_DST, MANIFEST_VERSIONED_DST])
   await fs.promises.writeFile(dst, JSON.stringify(manifest, null, 2), "utf8");
-}
 
 for (const asset of ASSETS) {
   const srcPath = path.join(ROOT, asset);
@@ -29,8 +27,8 @@ for (const asset of ASSETS) {
   if (fs.existsSync(srcPath)) {
     await fs.promises.cp(srcPath, destPath, {
       dereference: true,
-      recursive: true,
       force: true,
+      recursive: true,
       filter: src => {
         const fileName = path.basename(src);
         if (fileName === "macros") return false;
@@ -54,9 +52,7 @@ async function removeEmptyDirs(dir) {
     await Promise.all(files.map(file => removeEmptyDirs(path.join(dir, file))));
     files = await fs.promises.readdir(dir);
   }
-  if (files.length === 0) {
-    await fs.promises.rmdir(dir);
-  }
+  if (files.length === 0) await fs.promises.rmdir(dir);
 }
 
 await removeEmptyDirs(SYSTEM_DIR);
