+++ src/macros/ability-automation/ward-use-ability.mjs
@@ -1,5 +1,3 @@
 const isAttack = !!scope?.execution?.isAttack;
-const thisArmament = !!(scope?.execution?.armament?.uuid === scope?.item?.uuid);
-if (isAttack && thisArmament) {
-  scope.effect?.delete();
-}
+const thisArmament = scope?.execution?.armament?.uuid === scope?.item?.uuid;
+if (isAttack && thisArmament) scope.effect?.delete();
