+++ src/macros/gm-utilities/clean-default-item-pile.mjs
@@ -1,7 +1,2 @@
 const defaultItemPile = game.actors.getName("Default Item Pile");
-if (defaultItemPile) {
-  await defaultItemPile.deleteEmbeddedDocuments(
-    "Item",
-    defaultItemPile.items.map(i => i.id),
-  );
-}
+if (defaultItemPile) await defaultItemPile.deleteEmbeddedDocuments("Item", defaultItemPile.items.map(i => i.id));
