+++ src/module/applications/sheets/mixins/equipment-drop-sheet-mixin.mjs
@@ -29,53 +29,41 @@ export default function EquipmentDropSheetMixin(Base) {
           return;
         }
 
-        if (doc.type === "equipment" && doc.actor && !doc.isOwner) {
-          return;
-        }
+        if (doc.type === "equipment" && doc.actor && !doc.isOwner) return;
 
         // Create reference to old elder
         const _elder = doc.elder;
 
         // Optionally stack equipment
         let stack = false;
-        if (doc.type === "equipment") {
-          stack = await this._stackEquipment(doc);
-        }
+        if (doc.type === "equipment") stack = await this._stackEquipment(doc);
 
         // Handle moving equipment around within inventory instead of duplicating it
         if (
-          doc.type === "equipment" &&
-          doc.actor &&
-          (_elder?.type === "equipment" || _elder?.documentName === "Actor") &&
-          this.document.actor &&
-          doc.checkAncestor(this.document.actor)
+          doc.type === "equipment"
+          && doc.actor
+          && (_elder?.type === "equipment" || _elder?.documentName === "Actor")
+          && this.document.actor
+          && doc.checkAncestor(this.document.actor)
         ) {
           const oldElder = await resolveDocument(_elder);
           if (stack) {
             await doc.delete();
-            if (oldElder) {
-              await oldElder.sheet?.render();
-            }
+            if (oldElder) await oldElder.sheet?.render();
             return stack;
           } else if (this.document.type === "equipment" && this.document.system.storage.enabled) {
             await doc.update({ "system._sup": this.document.id });
-            if (oldElder) {
-              await oldElder.sheet?.render();
-            }
+            if (oldElder) await oldElder.sheet?.render();
             return doc;
           } else if (this.document.documentName === "Actor") {
             await doc.update({ "system._sup": null });
-            if (oldElder) {
-              await oldElder.sheet?.render();
-            }
+            if (oldElder) await oldElder.sheet?.render();
             return doc;
           }
         }
 
         if (stack) {
-          if (doc.isOwner) {
-            await doc.delete();
-          }
+          if (doc.isOwner) await doc.delete();
           return stack;
         }
 
@@ -85,9 +73,8 @@ export default function EquipmentDropSheetMixin(Base) {
         // Delete old equipment if it was moved instead of duplicated
         if (created?.type === "equipment") {
           const elder = await doc.getElder();
-          if (doc.isOwner && (elder?.type === "equipment" || elder?.documentName === "Actor") && doc.uuid) {
+          if (doc.isOwner && (elder?.type === "equipment" || elder?.documentName === "Actor") && doc.uuid)
             await doc.delete();
-          }
         }
         return created;
       }
